Enum guppy::graph::cargo::InitialsPlatform

pub enum InitialsPlatform {
    Host,
    Standard,
    ProcMacrosOnTarget,
}

For a given Cargo build simulation, what platform to assume the initials are being built on.

Variants

Host

Assume that the initials are being built on the host platform.

This is most useful for “continuing” simulations, where it is already known that some packages are being built on the host and one wishes to find their dependencies.

Standard

Assume a standard build.

In this mode, all initials other than proc-macros are built on the target platform. Proc- macros, being compiler plugins, are built on the host.

This is the default for InitialsPlatform.

ProcMacrosOnTarget

Perform a standard build, and also build proc-macros on the target.

Proc-macro crates may include tests, which are run on the target platform. This option is most useful for such situations.