Trait guppy::graph::cargo::CargoPostfilter

pub trait CargoPostfilter<'g> {
    fn accept_package(
        &mut self, 
        phase: CargoResolvePhase<'g, '_>, 
        link: PackageLink<'g>
    ) -> bool;

    fn accept_feature(
        &mut self, 
        phase: CargoResolvePhase<'g, '_>, 
        link: CrossLink<'g>
    ) -> bool { ... }
}

A final filter for if a package or feature link should be traversed during Cargo resolution.

This may be useful for more advanced control over package and feature resolution.

Required methods

fn accept_package(
&mut self,
phase: CargoResolvePhase<'g, '_>,
link: PackageLink<'g>
) -> bool

Returns true if this package link should be considered during a resolve operation.

This is called for the V1 and V1Install resolvers, and as part of the default implementation for accept_feature.

Returning false does not prevent the to package from being included if it's reachable through other means.

Loading content...
Provided methods

fn accept_feature(
&mut self,
phase: CargoResolvePhase<'g, '_>,
link: CrossLink<'g>
) -> bool

Returns true if this feature link should be considered during a resolve operation.

This is called for all resolvers. The default implementation forwards to accept_package.

Returning false does not prevent the to package from being included if it's reachable through other means.

The provided implementation forwards to accept_package. It is possible to customize this if necessary.

Loading content...