Enum guppy::graph::ExternalSource

#[non_exhaustive]
pub enum ExternalSource<'g> {
    Registry(&'g str),
    Git {
        repository: &'g str,
        req: GitReq<'g>,
        resolved: &'g str,
    },
}

More information about an external source.

This provides information about whether an external dependency is a Git dependency or fetched from a registry.

Returned by PackageSource::parse_external.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Registry(&'g str)

This is a registry source, e.g. "registry+https://github.com/rust-lang/crates.io-index".

The associated data is the part of the string after the initial "registry+".

Examples

use guppy::graph::ExternalSource;

let source = "registry+https://github.com/rust-lang/crates.io-index";
let parsed = ExternalSource::new(source).expect("this source is understood by guppy");

assert_eq!(
    parsed,
    ExternalSource::Registry("https://github.com/rust-lang/crates.io-index"),
);

Git

Fields

repository: &'g str

The repository for this Git source. For the above example, this would be "https://github.com/rust-lang/cargo.git".

req: GitReq<'g>

The revision requested in Cargo.toml. This may be a tag, a branch or a specific revision (commit hash).

For the above example, req would be GitSource::Branch("main").

resolved: &'g str

The resolved revision, as specified in Cargo.lock.

For the above example, resolved_hash would be "0227f048fcb7c798026ede6cc20c92befc84c3a4".

This is always a commit hash, and if req is GitReq::Rev then it is expected to be the same hash. (However, this is not verified by guppy.)

This is a Git source.

An example of a Git source string is "git+https://github.com/rust-lang/cargo.git?branch=main#0227f048fcb7c798026ede6cc20c92befc84c3a4". In this case, the Cargo.toml would have contained:

cargo = { git = "https://github.com/rust-lang/cargo.git", branch = "main" }

and the Cargo.lock would have contained:

[[package]]
name = "cargo"
version = "0.46.0"
source = "git+https://github.com/rust-lang/cargo.git?branch=main#0227f048fcb7c798026ede6cc20c92befc84c3a4
dependencies = [ ... ]

For more, see Specifying dependencies from git repositories in the Cargo book.

Examples

use guppy::graph::{ExternalSource, GitReq};

// A branch source.
let source = "git+https://github.com/rust-lang/cargo.git?branch=main#0227f048fcb7c798026ede6cc20c92befc84c3a4";
let parsed = ExternalSource::new(source).expect("this source is understood by guppy");

assert_eq!(
    parsed,
    ExternalSource::Git {
        repository: "https://github.com/rust-lang/cargo.git",
        req: GitReq::Branch("main"),
        resolved: "0227f048fcb7c798026ede6cc20c92befc84c3a4",
    }
);

// A tag source.
let source = "git+https://github.com/rust-lang/cargo.git?tag=v0.46.0#0227f048fcb7c798026ede6cc20c92befc84c3a4";
let parsed = ExternalSource::new(source).expect("this source is understood by guppy");

assert_eq!(
    parsed,
    ExternalSource::Git {
        repository: "https://github.com/rust-lang/cargo.git",
        req: GitReq::Tag("v0.46.0"),
        resolved: "0227f048fcb7c798026ede6cc20c92befc84c3a4",
    }
);

// A revision source.
let source = "git+https://github.com/rust-lang/cargo.git?rev=0227f048fcb7c798026ede6cc20c92befc84c3a4#0227f048fcb7c798026ede6cc20c92befc84c3a4";
let parsed = ExternalSource::new(source).expect("this source is understood by guppy");

assert_eq!(
    parsed,
    ExternalSource::Git {
        repository: "https://github.com/rust-lang/cargo.git",
        req: GitReq::Rev("0227f048fcb7c798026ede6cc20c92befc84c3a4"),
        resolved: "0227f048fcb7c798026ede6cc20c92befc84c3a4",
    }
);

// A default source.
let source = "git+https://github.com/gyscos/zstd-rs.git#bc874a57298bdb500cdb5aeac5f23878b6480d0b";
let parsed = ExternalSource::new(source).expect("this source is understood by guppy");

assert_eq!(
    parsed,
    ExternalSource::Git {
        repository: "https://github.com/gyscos/zstd-rs.git",
        req: GitReq::Default,
        resolved: "bc874a57298bdb500cdb5aeac5f23878b6480d0b",
    }
);

Implementations

impl<'g> ExternalSource<'g>

pub const REGISTRY_PLUS: &'static str = "registry+"

The string "registry+".

Used for matching with the Registry variant.

pub const GIT_PLUS: &'static str = "git+"

The string "git+".

Used for matching with the Git variant.

pub const BRANCH_EQ: &'static str = "?branch="

The string "?branch=".

Used for matching with the Git variant.

pub const TAG_EQ: &'static str = "?tag="

The string "?tag=".

Used for matching with the Git variant.

pub const REV_EQ: &'static str = "?rev="

The string "?rev=".

Used for matching with the Git variant.

pub const CRATES_IO_URL: &'static str = "https://github.com/rust-lang/crates.io-index"

The URL for the crates.io registry.

This lacks the leading "registry+“ that’s part of the PackageSource.

pub fn new(source: &'g str) -> Option<Self>

Attempts to parse the given string as an external source.

Returns None if the string could not be recognized as an external source.