guppy::graph
Enum PackageSource

pub enum PackageSource<'g> {
    Workspace(&'g Utf8Path),
    Path(&'g Utf8Path),
    External(&'g str),
}

The source of a package.

This enum contains information about where a package is found, and whether it is inside or outside the workspace.

Variants

Workspace(&'g Utf8Path)

This package is in the workspace.

The path is relative to the workspace root.

Path(&'g Utf8Path)

This package is a path dependency that isn’t in the workspace.

The path is relative to the workspace root.

External(&'g str)

This package is an external dependency.

For packages retrieved from crates.io, the source is the string "registry+https://github.com/rust-lang/crates.io-index".

For packages retrieved from other registries, the source begins with "registry+".

For packages retrieved from Git repositories, the source begins with "git+".

Implementations

impl<'g> PackageSource<'g>

pub const CRATES_IO_REGISTRY: &'static str = "registry+https://github.com/rust-lang/crates.io-index"

The path to the crates.io registry.

pub fn is_workspace(&self) -> bool

Returns true if this package source represents a workspace.

pub fn is_path(&self) -> bool

Returns true if this package source represents a path dependency that isn’t in the workspace.

pub fn is_external(&self) -> bool

Returns true if this package source represents an external dependency.

pub fn is_crates_io(&self) -> bool

Returns true if the source is crates.io.

pub fn is_local(&self) -> bool

Returns true if this package is a local dependency, i.e. either in the workspace or a local path.

pub fn workspace_path(&self) -> Option<&'g Utf8Path>

Returns the path if this is a workspace dependency, or None if this is a non-workspace dependency.

The path is relative to the workspace root.

pub fn local_path(&self) -> Option<&'g Utf8Path>

Returns the local path if this is a local dependency, or None if it is an external dependency.

The path is relative to the workspace root.

pub fn external_source(&self) -> Option<&'g str>

Returns the external source if this is an external dependency, or None if it is a local dependency.

pub fn parse_external(&self) -> Option<ExternalSource<'g>>

Attempts to parse an external source.

Returns None if the external dependency could not be recognized, or if it is a local dependency.

For more about external sources, see the documentation for ExternalSource.

impl PackageSource<'_>

pub fn to_summary_source(&self) -> SummarySource

Available on crate feature summaries only.

Converts a PackageSource into a SummarySource.

Requires the summaries feature to be enabled.