guppy::graph::feature
Module feature_list

A sorted, deduplicated list of features from a single package.

Structs

DisplayFeatures

A pretty-printer for a list of features.

FeatureList

A sorted, deduplicated list of features from a single package.

IntoIter

An owned iterator over a FeatureList.

Iter

A borrowed iterator over a FeatureList.