Struct guppy::graph::feature::feature_list::FeatureList

pub struct FeatureList<'g> { /* private fields */ }

A sorted, deduplicated list of features from a single package.

This provides a convenient way to query and print out lists of features.

Returned by methods on FeatureSet.

Implementations

impl<'g> FeatureList<'g>

pub fn new( package: PackageMetadata<'g>, labels: impl IntoIterator<Item = FeatureLabel<'g>> ) -> Self

Creates a new FeatureList from a package and an iterator over feature labels.

pub fn package(&self) -> &PackageMetadata<'g>

Returns the package corresponding to this feature list.

pub fn contains(&self, label: FeatureLabel<'_>) -> bool

Returns true if this feature list contains this feature label.

pub fn has_base(&self) -> bool

Returns true if this feature list contains the “base” feature.

The “base” feature represents the package with no features enabled.

pub fn has_named_feature(&self, feature_name: &str) -> bool

Returns true if this feature list contains the specified named feature.

pub fn has_optional_dependency(&self, dep_name: &str) -> bool

Returns true if this feature list contains the specified optional dependency.

pub fn labels(&self) -> &[FeatureLabel<'g>]

Returns the list of labels as a slice.

This slice is guaranteed to be sorted and unique.

pub fn named_features(&self) -> impl Iterator<Item = &'g str> + '_

Returns an iterator containing all named features.

The iterator is guaranteed to be sorted and unique.

pub fn optional_deps(&self) -> impl Iterator<Item = &'g str> + '_

Returns an iterator containing all optional dependencies.

The iterator is guaranteed to be sorted and unique.

pub fn iter<'a>(&'a self) -> Iter<'g, 'a>

Returns a borrowed iterator over feature IDs.

pub fn display_features<'a>(&'a self) -> DisplayFeatures<'g, 'a>

Returns a pretty-printer over the list of feature labels.

pub fn into_labels(self) -> Vec<FeatureLabel<'g>>

Returns a vector of feature labels.

The vector is guaranteed to be sorted and unique.