Function guppy::graph::feature::feature_filter

pub fn feature_filter<'g: 'a, 'a>(
    base: impl FeatureFilter<'g> + 'a, 
    features: impl IntoIterator<Item = &'a str>
) -> impl FeatureFilter<'g> + 'a

Returns a FeatureFilter that selects everything from the base filter, plus these additional feature names -- regardless of what package they are in.

This is equivalent to a build with --features, and is typically meant to be used with one package.

For filtering by feature IDs, use feature_id_filter.