Struct guppy::graph::BuildTarget

pub struct BuildTarget<'g> { /* private fields */ }

A build target in a package.

A build target consists of one or more source files which can be compiled into a crate.

For more, see Cargo Targets in the Cargo reference.

Implementations

impl<'g> BuildTarget<'g>

pub fn id(&self) -> BuildTargetId<'g>

Returns the unique identifier for this build target.

pub fn name(&self) -> &'g str

Returns the name of this build target.

pub fn kind(&self) -> BuildTargetKind<'g>

Returns the kind of this build target.

pub fn required_features(&self) -> &'g [String]

Returns the features required for this build target.

This setting has no effect on the library target.

For more, see The required-features field in the Cargo reference.

pub fn path(&self) -> &'g Utf8Path

Returns the absolute path of the location where the source for this build target is located.

pub fn edition(&self) -> &'g str

Returns the Rust edition for this build target.

pub fn doc_tests(&self) -> bool

Returns true if documentation tests are enabled for this build target.