Struct guppy::graph::DependencyReq

pub struct DependencyReq<'g> { /* private fields */ }

Information about a specific kind of dependency (normal, build or dev) from a package to another package.

Usually found within the context of a PackageLink.

Implementations

impl<'g> DependencyReq<'g>

pub fn is_present(&self) -> bool

Returns true if there is at least one Cargo.toml entry corresponding to this requirement.

For example, if this dependency is specified in the [dev-dependencies] section, edge.dev().is_present() will return true.

pub fn status(&self) -> EnabledStatus<'g>

Returns the enabled status of this dependency.

status is the union of default_features and no_default_features.

See the documentation for EnabledStatus for more.

pub fn default_features(&self) -> EnabledStatus<'g>

Returns the enabled status of this dependency when default-features = true.

See the documentation for EnabledStatus for more.

pub fn no_default_features(&self) -> EnabledStatus<'g>

Returns the enabled status of this dependency when default-features = false.

This is generally less useful than status or default_features, but is provided for completeness.

See the documentation for EnabledStatus for more.

pub fn features(&self) -> impl Iterator<Item = &'g str>

Returns a list of all features possibly enabled by this dependency. This includes features that are only turned on if the dependency is optional, or features enabled by inactive platforms.

pub fn feature_status(&self, feature: &str) -> EnabledStatus<'g>

Returns the enabled status of this feature.

Note that as of Rust 1.42, the default feature resolver behaves in potentially surprising ways. See the Cargo reference for more.

See the documentation for EnabledStatus for more.