Struct guppy::graph::PackageMetadata

pub struct PackageMetadata<'g> { /* private fields */ }

Information about a specific package in a PackageGraph.

Most of the metadata is extracted from Cargo.toml files. See the Cargo.toml reference for more details.

Implementations

impl<'g> PackageMetadata<'g>

pub fn id(&self) -> &'g PackageId

Returns the unique identifier for this package.

pub fn graph(&self) -> &'g PackageGraph

Returns the package graph this PackageMetadata is derived from.

pub fn to_package_query( &self, direction: DependencyDirection ) -> PackageQuery<'g>

Creates a PackageQuery consisting of this package, in the given direction.

The PackageQuery can be used to inspect dependencies in this graph.

pub fn to_package_set(&self) -> PackageSet<'g>

Creates a PackageSet consisting of just this package.

pub fn to_feature_set(&self, features: impl FeatureFilter<'g>) -> FeatureSet<'g>

Creates a FeatureSet that consists of all features in the package that match the given named filter.

pub fn direct_links_directed( &self, direction: DependencyDirection ) -> impl Iterator<Item = PackageLink<'g>> + 'g

Returns PackageLink instances corresponding to the direct dependencies for this package in the specified direction.

pub fn direct_links(&self) -> impl Iterator<Item = PackageLink<'g>> + 'g

Returns PackageLink instances corresponding to the direct dependencies for this package.

pub fn reverse_direct_links(&self) -> impl Iterator<Item = PackageLink<'g>> + 'g

Returns PackageLink instances corresponding to the packages that directly depend on this one.

pub fn link_between( &self, other: &PackageId, direction: DependencyDirection ) -> Result<Option<PackageLink<'g>>, Error>

Returns the direct PackageLink between self and other in the specified direction:

Forward: from self to other

Reverse: from other to self

Returns None if the direct link does not exist, or an error if to isn’t found in self.graph().

pub fn link_to(&self, to: &PackageId) -> Result<Option<PackageLink<'g>>, Error>

Returns the direct PackageLink from self to the specified package, or None if self does not directly depend on the specified package.

Returns an error if to isn’t found in self.graph().

pub fn link_from( &self, from: &PackageId ) -> Result<Option<PackageLink<'g>>, Error>

Returns the direct PackageLink from the specified package to self, or None if the specified package does not directly depend on self.

Returns an error if from isn’t found in self.graph().

pub fn name(&self) -> &'g str

Returns the name of this package.

This is the same as the name field of Cargo.toml.

pub fn version(&self) -> &'g Version

Returns the version of this package as resolved by Cargo.

This is the same as the version field of Cargo.toml.

pub fn authors(&self) -> &'g [String]

Returns the authors of this package.

This is the same as the authors field of Cargo.toml.

pub fn description(&self) -> Option<&'g str>

Returns a short description for this package.

This is the same as the description field of Cargo.toml.

pub fn license(&self) -> Option<&'g str>

Returns an SPDX 2.1 license expression for this package, if specified.

This is the same as the license field of Cargo.toml. Note that guppy does not perform any validation on this, though crates.io does if a crate is uploaded there.

pub fn license_file(&self) -> Option<&'g Utf8Path>

Returns the path to a license file for this package, if specified.

This is the same as the license_file field of Cargo.toml. It is typically only specified for nonstandard licenses.

pub fn source(&self) -> PackageSource<'g>

Returns the source from which this package was retrieved.

This may be the workspace path, an external path, or a registry like crates.io.

pub fn in_workspace(&self) -> bool

Returns true if this package is in the workspace.

For more detailed information, use source().

pub fn manifest_path(&self) -> &'g Utf8Path

Returns the full path to the Cargo.toml for this package.

This is specific to the system that cargo metadata was run on.

pub fn categories(&self) -> &'g [String]

Returns categories for this package.

This is the same as the categories field of Cargo.toml. For packages on crates.io, returned values are guaranteed to be valid category slugs.

pub fn keywords(&self) -> &'g [String]

Returns keywords for this package.

This is the same as the keywords field of Cargo.toml.

pub fn readme(&self) -> Option<&'g Utf8Path>

Returns a path to the README for this package, if specified.

This is the same as the readme field of Cargo.toml. The path returned is relative to the directory the Cargo.toml is in (i.e. relative to the parent of self.manifest_path()).

pub fn repository(&self) -> Option<&'g str>

Returns the source code repository for this package, if specified.

This is the same as the repository field of Cargo.toml.

pub fn homepage(&self) -> Option<&'g str>

Returns the homepage for this package, if specified.

This is the same as the homepage field of Cargo.toml.

pub fn documentation(&self) -> Option<&'g str>

Returns the documentation URL for this package, if specified.

This is the same as the homepage field of Cargo.toml.

pub fn edition(&self) -> &'g str

Returns the Rust edition this package is written against.

This is the same as the edition field of Cargo.toml. It is "2015" by default.

pub fn metadata_table(&self) -> &'g JsonValue

Returns the freeform metadata table for this package.

This is the same as the package.metadata section of Cargo.toml. This section is typically used by tools which would like to store package configuration in Cargo.toml.

pub fn links(&self) -> Option<&'g str>

Returns the name of a native library this package links to, if specified.

This is the same as the links field of Cargo.toml. See The links Manifest Key in the Cargo book for more details.

pub fn publish(&self) -> PackagePublish<'g>

Returns the registries to which this package may be published.

This is derived from the publish field of Cargo.toml.

pub fn default_run(&self) -> Option<&'g str>

Returns the binary that is run by default, if specified.

Information about this binary can be queried using the build_target method.

This is derived from the default-run field of Cargo.toml.

pub fn minimum_rust_version(&self) -> Option<&'g Version>

Returns the minimum Rust compiler version, which should be able to compile the package, if specified.

This is the same as the rust-version field of Cargo.toml. For more, see the rust-version field in the Cargo reference.

pub fn rust_version(&self) -> Option<&'g VersionReq>

👎Deprecated since 0.17.1: use Self::rust_version instead, it returns a Version

Returns the minimum Rust compiler version, which should be able to compile the package, if specified.

Returned as a semver::VersionReq. This is actually not correct – it is deprecated and will go away in the next major version of guppy: use Self::minimum_rust_version instead.

This is the same as the rust-version field of Cargo.toml. For more, see the rust-version field in the Cargo reference.

pub fn build_targets(&self) -> impl Iterator<Item = BuildTarget<'g>>

Returns all the build targets for this package.

For more, see Cargo Targets in the Cargo reference.

pub fn build_target(&self, id: &BuildTargetId<'_>) -> Option<BuildTarget<'g>>

Looks up a build target by identifier.

pub fn is_proc_macro(&self) -> bool

Returns true if this package is a procedural macro.

For more about procedural macros, see Procedural Macros in the Rust reference.

pub fn has_build_script(&self) -> bool

Returns true if this package has a build script.

Cargo only follows build dependencies if a build script is set.

For more about build scripts, see Build Scripts in the Cargo reference.

pub fn has_default_feature(&self) -> bool

Returns true if this package has a named feature named default.

For more about default features, see The [features] section in the Cargo reference.

pub fn default_feature_id(&self) -> FeatureId<'g>

Returns the FeatureId corresponding to the default feature.

pub fn named_features(&self) -> impl Iterator<Item = &'g str> + 'g

Returns the list of named features available for this package. This will include a feature named “default” if it is defined.

A named feature is listed in the [features] section of Cargo.toml. For more, see the reference.