Struct guppy::graph::Workspace

pub struct Workspace<'g> { /* private fields */ }

Information about a workspace, parsed from metadata returned by cargo metadata.

For more about workspaces, see Cargo Workspaces in The Rust Programming Language.

Implementations

impl<'g> Workspace<'g>

pub fn root(&self) -> &'g Utf8Path

Returns the workspace root.

pub fn target_directory(&self) -> &'g Utf8Path

Returns the target directory in which output artifacts are stored.

pub fn member_count(&self) -> usize

Returns the number of packages in this workspace.

pub fn contains_name(&self, name: impl AsRef<str>) -> bool

Returns true if the workspace contains a package by the given name.

pub fn contains_path(&self, path: impl AsRef<Utf8Path>) -> bool

Returns true if the workspace contains a package by the given workspace path.

pub fn iter( &self ) -> impl Iterator<Item = PackageMetadata<'g>> + ExactSizeIterator

Returns an iterator over package metadatas, sorted by the path they’re in.

pub fn iter_by_path( &self ) -> impl Iterator<Item = (&'g Utf8Path, PackageMetadata<'g>)> + ExactSizeIterator

Returns an iterator over workspace paths and package metadatas, sorted by the path they’re in.

pub fn iter_by_name( &self ) -> impl Iterator<Item = (&'g str, PackageMetadata<'g>)> + ExactSizeIterator

Returns an iterator over workspace names and package metadatas, sorted by names.

pub fn member_ids( &self ) -> impl Iterator<Item = &'g PackageId> + ExactSizeIterator

Returns an iterator over package IDs for workspace members. The package IDs will be returned in the same order as members, sorted by the path they’re in.

pub fn member_by_path( &self, path: impl AsRef<Utf8Path> ) -> Result<PackageMetadata<'g>, Error>

Maps the given path to the corresponding workspace member.

Returns an error if the path didn’t match any workspace members.

pub fn members_by_paths<B>( &self, paths: impl IntoIterator<Item = impl AsRef<Utf8Path>> ) -> Result<B, Error>
where B: FromIterator<PackageMetadata<'g>>,

Maps the given paths to their corresponding workspace members, returning a new value of the specified collection type (e.g. Vec).

Returns an error if any of the paths were unknown.

pub fn member_by_name( &self, name: impl AsRef<str> ) -> Result<PackageMetadata<'g>, Error>

Maps the given name to the corresponding workspace member.

Returns an error if the name didn’t match any workspace members.

pub fn members_by_names<B>( &self, names: impl IntoIterator<Item = impl AsRef<str>> ) -> Result<B, Error>
where B: FromIterator<PackageMetadata<'g>>,

Maps the given names to their corresponding workspace members, returning a new value of the specified collection type (e.g. Vec).

Returns an error if any of the paths were unknown.

pub fn metadata_table(&self) -> &'g JsonValue

Returns the freeform metadata table for this workspace.

This is the same as the workspace.metadata section of Cargo.toml. This section is typically used by tools which would like to store workspace configuration in Cargo.toml.