guppy::graph
Module summaries

Available on crate feature summaries only.

Generate build summaries from CargoSet instances.

Requires the summaries feature to be enabled.

Modules

diff

Compare and diff summaries.

Structs

CargoOptionsSummary

A summary of Cargo options used to build a CargoSet.

FeaturesOnlySummary

Summary information for a features-only package.

PackageInfo

Information about a package in a summary that isn’t part of the unique identifier.

PackageSetSummary

A set of packages specified in a summary. Can be resolved into a PackageSet given a PackageGraph.

Summary

An in-memory representation of a build summary.

SummaryId

A unique identifier for a package in a build summary.

ThirdPartySummary

A selector for external, third-party packages.

Enums

GitReqSummary

A summary specification for a Git branch, tag or revision.

InitialsPlatformSummary

Summary information for InitialsPlatform.

PackageStatus

The status of a package in a summary, such as whether it is part of the initial build set.

SummarySource

The location of a package.

ThirdPartySource

Describes locations where non-workspace packages (path or external) can be found.

Functions

serialize_forward_slashes

Serializes a path with forward slashes on Windows.

Type Aliases

PackageMap

A type representing a package map as used in Summary instances.