Enum guppy::graph::summaries::diff::SummaryDiffStatus

pub enum SummaryDiffStatus<'a> {
    Added {
        info: &'a PackageInfo,
    },
    Removed {
        old_info: &'a PackageInfo,
    },
    Modified {
        old_version: Option<&'a Version>,
        old_source: Option<&'a SummarySource>,
        old_status: Option<PackageStatus>,
        new_status: PackageStatus,
        added_features: BTreeSet<&'a str>,
        removed_features: BTreeSet<&'a str>,
        unchanged_features: BTreeSet<&'a str>,
    },
}

This is supported on crate feature summaries only.

The diff status for a particular summary ID and source.

Variants

Added

Fields

info: &'a PackageInfo

The information for this package.

This package was added.

Removed

Fields

old_info: &'a PackageInfo

The information this package used to have.

This package was removed.

Modified

Fields

old_version: Option<&'a Version>

The old version of this package, if the version changed.

old_source: Option<&'a SummarySource>

The old source of this package, if the source changed.

old_status: Option<PackageStatus>

The old status of this package, if the status changed.

new_status: PackageStatus

The current status of this package.

added_features: BTreeSet<&'a str>

The set of features added to the package.

removed_features: BTreeSet<&'a str>

The set of features removed from the package.

unchanged_features: BTreeSet<&'a str>

The set of features which were enabled both in both the old and new summaries.

Some details about the package changed:

a feature was added or removed

the version or source changed.

Implementations

impl<'a> SummaryDiffStatus<'a>

pub fn tag(&self) -> SummaryDiffTag

Returns the tag for this status.

The tag is similar to this enum, except it has no associated data.

pub fn latest_status(&self) -> PackageStatus

Returns the new package status if available, otherwise the old status.