guppy::graph::summaries::diff
Struct SummaryDiff

pub struct SummaryDiff<'a> {
    pub target_packages: PackageDiff<'a>,
    pub host_packages: PackageDiff<'a>,
}

Available on crate feature summaries only.

A diff of two package summaries.

This struct contains information on the packages that were changed, as well as those that were not.

Human-readable reports

The report method can be used with fmt::Display to generate a friendly, human-readable report.

Machine-readable serialization

A SummaryDiff can be serialized through serde. The output format is part of the API.

An example of TOML-serialized output:

[[target-packages.changed]]
name = "dep"
version = "0.4.3"
crates-io = true
change = "added"
status = "direct"
features = ["std"]

[[target-packages.changed]]
name = "foo"
version = "1.2.3"
workspace-path = "foo"
change = "modified"
new-status = "initial"
added-features = ["feature2"]
removed-features = []
unchanged-features = ["default", "feature1"]

[[target-packages.unchanged]]
name = "no-changes"
version = "1.5.3"
crates-io = true
status = "transitive"
features = ["default"]

[[host-packages.changed]]
name = "dep"
version = "0.4.2"
crates-io = true
change = "removed"
old-status = "direct"
old-features = ["std"]

Fields
target_packages: PackageDiff<'a>

Diff of target packages.

host_packages: PackageDiff<'a>

Diff of host packages.

Implementations

impl<'a> SummaryDiff<'a>

pub fn new(old: &'a Summary, new: &'a Summary) -> SummaryDiff<'a>

Computes a diff between two summaries.

pub fn is_changed(&self) -> bool

Returns true if there are any changes in this diff.

pub fn is_unchanged(&self) -> bool

Returns true if there are no changes in this diff.

pub fn report<'b>(&'b self) -> SummaryReport<'a, 'b>

Returns a report for this diff.

This report can be used with fmt::Display.