guppy::graph::summaries
Enum PackageStatus

pub enum PackageStatus {
    Initial,
    Workspace,
    Direct,
    Transitive,
}

Available on crate feature summaries only.

The status of a package in a summary, such as whether it is part of the initial build set.

The ordering here determines what order packages will be written out in the summary.

Variants

Initial

This package is part of the requested build set.

Workspace

This is a workspace package that isn’t part of the requested build set.

Direct

This package is a direct non-workspace dependency.

A Direct package may also be transitively included.

Transitive

This package is a transitive non-workspace dependency.