Enum guppy::graph::summaries::SummarySource

pub enum SummarySource {
    Workspace {
        workspace_path: PathBuf,
    },
    Path {
        path: PathBuf,
    },
    CratesIo,
    External {
        source: String,
    },
}

The location of a package.

Variants

Workspace

A workspace path.

Fields of Workspace

workspace_path: PathBuf

The path of this package, relative to the workspace root.

Path

A non-workspace path.

The path is expected to be relative to the workspace root.

Fields of Path

path: PathBuf

The path of this package, relative to the workspace root.

CratesIo

The crates.io registry.

External

An external source that's not the crates.io registry, such as an alternate registry or a git repository.

Fields of External

source: String

The external source.

Implementations

impl SummarySource

pub fn workspace(workspace_path: impl Into<PathBuf>) -> SummarySource

Creates a new SummarySource representing a workspace source.

pub fn path(path: impl Into<PathBuf>) -> SummarySource

Creates a new SummarySource representing a non-workspace path source.

pub fn crates_io() -> SummarySource

Creates a new SummarySource representing the crates.io registry.

pub fn external(source: impl Into<String>) -> SummarySource

Creates a new SummarySource representing an external source like a Git repository or a custom registry.