Struct guppy::Platform

pub struct Platform<'a> { /* fields omitted */ }

A platform to evaluate target specs against.

Implementations

impl<'a> Platform<'a>

pub fn new(
triple: impl AsRef<str>,
target_features: TargetFeatures<'a>
) -> Result<Platform<'a>, Error>

Creates a new Platform from the given built-in triple and target features.

Returns None if this platform wasn't known to target-spec.

pub fn custom(
target_info: &'a TargetInfo<'a>,
target_features: TargetFeatures<'a>
) -> Platform<'a>

Creates a new, custom platform from a TargetInfo and target features.

Custom platforms are often found in embedded and similar environments. For built-in platforms, new is recommended instead.

pub fn add_flags(&mut self, flags: impl IntoIterator<Item = &'a str>)

Adds a set of flags to accept.

A flag is a single token like the foo in cfg(not(foo)).

A default cargo build will always evaluate flags to false, but custom wrappers may cause some flags to evaluate to true. For example, as of version 0.6, cargo web build will cause cargo_web to evaluate to true.

pub fn triple(&self) -> &'a str

Returns the target triple for this platform.

pub fn flags(&'b self) -> impl Iterator<Item = &'a str> + 'b

Returns the set of flags enabled for this platform.

pub fn has_flag(&self, flag: impl AsRef<str>) -> bool

Returns true if this flag was set with add_flags.

pub fn target_info(&self) -> &'a TargetInfo<'a>

Returns the underlying TargetInfo.

pub fn target_features(&self) -> &TargetFeatures<'a>

Returns the set of target features for this platform.

pub fn is_custom(&self) -> bool

Returns true if this is a custom platform, created by Platform::custom.

impl Platform<'static>

pub fn current() -> Option<Platform<'static>>

Returns the current platform, as detected at build time.

This will return None if the current platform was unknown to this version of target-spec.

impl<'a> Platform<'a>

Helpers for property testing

The methods in this section allow Platform instances to be used in property-based testing scenarios.

Currently, proptest 0.10 is supported if the proptest010 feature is enabled.

pub fn strategy(
target_features: impl Strategy<Value = TargetFeatures<'a>> + 'a
) -> impl Strategy<Value = Platform<'a>> + 'a

Given a way to generate TargetFeatures instances, this returns a Strategy that generates a platform at random.

Requires the proptest010 feature to be enabled.

Examples

use proptest::prelude::*;
use target_spec::{Platform, TargetFeatures};

// target_features is a strategy that always produces TargetFeatures::Unknown.
let target_features = Just(TargetFeatures::Unknown);
let strategy = Platform::strategy(target_features);

pub fn filtered_strategy(
triple_filter: impl Fn(&'static str) -> bool,
target_features: impl Strategy<Value = TargetFeatures<'a>> + 'a
) -> impl Strategy<Value = Platform<'a>> + 'a

A version of strategy that allows target triples to be filtered.

Requires the proptest010 feature to be enabled.