Module guppy::graph

Entry point for analyzing Cargo dependency graphs.

The main entry point for analyzing graphs is PackageGraph. See its documentation for more details.

Structs

DependencyEdge
Details about a specific dependency from a package to another package.

DependencyLink
Represents a dependency from one package to another.

DependencyMetadata
Information about a specific kind of dependency (normal, build or dev) from a package to another package.

DependsCache
An optional cache used to speed up depends_on queries.

DotWrite
A write target for Dot graphs. Use with the write! macro.

IntoIterIds
An iterator over package IDs in topological order.

IntoIterLinks
An iterator over dependency links.

IntoIterMetadatas
An iterator over package metadata in topological order.

PackageGraph
A graph of packages and dependencies between them, parsed from metadata returned by cargo metadata.

PackageGraphData
Per-package data for a PackageGraph instance.

PackageMetadata
Information about a specific package in a PackageGraph.

PackageSelect
A selector over a package graph.

Workspace
Information about a workspace, parsed from metadata returned by cargo metadata.

Enums

DependencyDirection
The direction in which to follow dependencies.

Traits

PackageDotVisitor
A visitor used for formatting dot graphs.