Module guppy::graph::cargo

Simulations of Cargo behavior.

Cargo comes with a set of algorithms to figure out what packages or features are built. This module reimplements those algorithms using guppy's data structures.

Structs

CargoOptions
Options for queries which simulate what Cargo does.

CargoSet
A set of packages and features, as would be built by Cargo.

Enums

CargoResolverVersion
The version of Cargo's feature resolver to use.