Struct guppy::graph::cargo::CargoSet

pub struct CargoSet<'g> { /* fields omitted */ }

A set of packages and features, as would be built by Cargo.

Cargo implements a set of algorithms to figure out which packages or features are built in a given situation. guppy implements those algorithms.

Implementations

impl<'g> CargoSet<'g>

pub fn new(query: FeatureQuery<'g>, opts: &CargoOptions) -> Result<Self, Error>

Creates a new CargoSet based on the given query and options.

This is also accessible through FeatureQuery::resolve_cargo(), and it may be more convenient to use that if the code is written in a "fluent" style.

pub fn feature_graph(&self) -> &FeatureGraph<'g>

Returns the feature graph for this CargoSet instance.

pub fn package_graph(&self) -> &'g PackageGraph

Returns the package graph for this CargoSet instance.

pub fn original_query(&self) -> &FeatureQuery<'g>

Returns the original query from which the CargoSet instance was constructed.

pub fn target_features(&self) -> &FeatureSet<'g>

Returns the feature set enabled on the target platform.

This represents the packages and features that are included as code in the final build artifacts. This is relevant for both cross-compilation and auditing.

pub fn host_features(&self) -> &FeatureSet<'g>

Returns the feature set enabled on the host platform.

This represents the packages and features that influence the final build artifacts, but whose code is generally not directly included.

This includes all procedural macros, including those specified in the initial query.

pub fn target_direct_deps(&self) -> &PackageSet<'g>

Returns the set of workspace and direct dependency packages on the target platform.

The packages in this set are a subset of the packages in target_features.

pub fn host_direct_deps(&self) -> &PackageSet<'g>

Returns the set of workspace and direct dependency packages on the host platform.

The packages in this set are a subset of the packages in host_features.

pub fn proc_macro_links<'a>(
&'a self
) -> impl Iterator<Item = PackageLink<'g>> + ExactSizeIterator + 'a

Returns PackageLink instances for procedural macro dependencies from target packages.

Procedural macros straddle the line between target and host: they're built for the host but generate code that is compiled for the target platform.

Notes

Procedural macro packages will be included in the host feature set.

The returned iterator will include proc macros that are depended on normally or in dev builds from initials (if include_dev is set), but not the ones in the [build-dependencies] section.

pub fn build_dep_links<'a>(
&'a self
) -> impl Iterator<Item = PackageLink<'g>> + ExactSizeIterator + 'a

Returns PackageLink instances for build dependencies from target packages.

Notes

For each link, the from is built on the target while the to is built on the host. It is possible (though rare) that a build dependency is also included as a normal dependency, or as a dev dependency in which case it will also be built on the target.

The returned iterators will not include build dependencies of host packages -- those are also built on the host.

impl<'g> CargoSet<'g>

pub fn to_summary(&self, opts: &CargoOptions) -> Result<Summary, Error>

Creates a build summary with the given options.

Requires the summaries feature to be enabled.