Module guppy::graph::feature

Graph analysis for individual features within a package.

FeatureGraph can be used to do a more precise analysis than is possible at the package level. For example, an optional feature not included a default build can potentially pull in a large number of extra dependencies. This module allows for those subgraphs to be filtered out.

Re-exports

pub use feature_list::FeatureList;

Modules

feature_list
A sorted, deduplicated list of features from a single package.

Structs

CrossLink
A feature dependency across packages.

Cycles
Contains information about dependency cycles in feature graphs.

FeatureFilterFn
A FeatureFilter which calls the function that's passed in.

FeatureGraph
A derived graph representing every feature of every package.

FeatureId
An identifier for a (package, feature) pair in a feature graph.

FeatureMetadata
Metadata for a feature within a package.

FeatureQuery
A query over a feature graph.

FeatureSet
A set of resolved feature IDs in a feature graph.

Enums

FeatureType
The type of a particular feature within a package.

Traits

FeatureFilter
Trait representing whether a feature within a package should be selected.

FeatureResolver
Represents whether a particular link within a feature graph should be followed during a resolve operation.

Functions

all_filter
Returns a FeatureFilter that selects all features from the given packages.

default_filter
Returns a FeatureFilter that selects default features from the given packages.

feature_filter
Returns a FeatureFilter that selects everything from the base filter, plus these additional feature names -- regardless of what package they are in.

feature_id_filter
Returns a FeatureFilter that selects everything from the base filter, plus some additional feature IDs.

none_filter
Returns a FeatureFilter that selects no features from the given packages.