Struct guppy::graph::feature::feature_list::FeatureList

pub struct FeatureList<'g> { /* private fields */ }

A sorted, deduplicated list of features from a single package.

This provides a convenient way to query and print out lists of features.

Returned by methods on FeatureSet.

Implementations

impl<'g> FeatureList<'g>

pub fn new(
package: PackageMetadata<'g>,
features: impl IntoIterator<Item = Option<&'g str>>
) -> Self

Creates a new FeatureList from a package and an iterator over features.

pub fn package(&self) -> &PackageMetadata<'g>

Returns the package corresponding to this feature list.

pub fn contains(&self, feature: impl AsRef<str>) -> bool

Returns true if this feature list contains this feature.

pub fn has_base(&self) -> bool

Returns true if this feature list contains the “base” feature.

The “base” feature represents the package with no features enabled.

pub fn features(&self) -> &[&'g str]

Returns the list of features as a slice.

This slice is guaranteed to be sorted and unique.

pub fn iter<'a>(&'a self) -> Iter<'g, 'a>

Returns a borrowed iterator over feature IDs.

pub fn display_features<'a>(&'a self) -> DisplayFeatures<'g, 'a>

Returns a pretty-printer over the list of features.

pub fn into_features(self) -> Vec<&'g str>

Returns a vector of feature names.

The vector is guaranteed to be sorted and unique.