Struct guppy::graph::feature::FeatureId

pub struct FeatureId<'g> { /* private fields */ }

An identifier for a (package, feature) pair in a feature graph.

Returned by various methods on FeatureGraph and FeatureQuery.

From impls are available for (&'g PackageId, &'g str) and (&'g PackageId, Option<&'g str>) tuples.

Implementations

impl<'g> FeatureId<'g>

pub fn new(package_id: &'g PackageId, label: FeatureLabel<'g>) -> Self

Creates a new FeatureId with the given PackageId and FeatureLabel.

pub fn named(package_id: &'g PackageId, feature_name: &'g str) -> Self

Creates a new FeatureId representing a named feature in the [features] section, or an implicit named feature created by an optional dependency.

pub fn optional_dependency(package_id: &'g PackageId, dep_name: &'g str) -> Self

Creates a new FeatureId representing an optional dependency.

pub fn base(package_id: &'g PackageId) -> Self

Creates a new FeatureId representing the base feature for a package.

pub fn package_id(&self) -> &'g PackageId

Returns the package ID.

pub fn label(&self) -> FeatureLabel<'g>

Returns the FeatureLabel associated with the feature.

pub fn is_base(&self) -> bool

Returns true if this is the base feature for the package.

pub fn is_optional_dependency(self) -> bool

Returns true if this is an optional dependency.

pub fn is_named(self) -> bool

Returns true if this is a named feature.