Struct guppy::graph::feature::FeatureSet

pub struct FeatureSet<'g> { /* fields omitted */ }

A set of resolved feature IDs in a feature graph.

Created by FeatureQuery::resolve, the FeatureGraph::resolve_ methods, or from PackageSet::to_feature_set.

Implementations

impl<'g> FeatureSet<'g>

pub fn graph(&self) -> &FeatureGraph<'g>

Returns the FeatureGraph that this feature set was computed against.

pub fn len(&self) -> usize

Returns the number of feature IDs in this set.

pub fn is_empty(&self) -> bool

Returns true if no feature IDs were resolved in this set.

pub fn contains<'a>(
&self,
feature_id: impl Into<FeatureId<'a>>
) -> Result<bool, Error>

Returns true if this set contains the given feature ID.

Returns an error if this feature ID was unknown.

pub fn union(&self, other: &Self) -> Self

Returns a FeatureSet that contains all packages present in at least one of self and other.

Panics

Panics if the package graphs associated with self and other don't match.

pub fn intersection(&self, other: &Self) -> Self

Returns a FeatureSet that contains all packages present in both self and other.

Panics

Panics if the package graphs associated with self and other don't match.

pub fn difference(&self, other: &Self) -> Self

Returns a FeatureSet that contains all packages present in self but not other.

Panics

Panics if the package graphs associated with self and other don't match.

pub fn symmetric_difference(&self, other: &Self) -> Self

Returns a FeatureSet that contains all packages present in exactly one of self and other.

Panics

Panics if the package graphs associated with self and other don't match.

pub fn features_for(
&self,
package_id: &PackageId
) -> Result<Option<FeatureList<'g>>, Error>

Returns a list of features present for this package, or None if this package is not present in the feature set.

Returns an error if the package ID was unknown.

pub fn to_package_set(&self) -> PackageSet<'g>

Converts this FeatureSet into a PackageSet containing all packages with any selected features (including the "base" feature).

pub fn feature_ids<'a>(
&'a self,
direction: DependencyDirection
) -> impl Iterator<Item = FeatureId<'g>> + ExactSizeIterator + 'a

Iterates over feature IDs, in topological order in the direction specified.

Cycles

The packages within a dependency cycle will be returned in arbitrary order, but overall topological order will be maintained.

pub fn features<'a>(
&'a self,
direction: DependencyDirection
) -> impl Iterator<Item = FeatureMetadata<'g>> + ExactSizeIterator + 'a

Iterates over feature metadatas, in topological order in the direction specified.

Cycles

The packages within a dependency cycle will be returned in arbitrary order, but overall topological order will be maintained.

pub fn packages_with_features<'a>(
&'a self,
direction: DependencyDirection
) -> impl Iterator<Item = FeatureList<'g>> + 'a

Iterates over package metadatas and their corresponding features, in topological order in the direction specified.

Cycles

The packages within a dependency cycle will be returned in arbitrary order, but overall topological order will be maintained.

pub fn root_ids<'a>(
&'a self,
direction: DependencyDirection
) -> impl Iterator<Item = FeatureId<'g>> + ExactSizeIterator + 'a

Returns the set of "root feature" IDs in the specified direction.

If direction is Forward, return the set of feature IDs that do not have any dependencies within the selected graph.

If direction is Reverse, return the set of feature IDs that do not have any dependents within the selected graph.

Cycles

If a root consists of a dependency cycle, all the packages in it will be returned in arbitrary order.

pub fn root_features<'a>(
&'a self,
direction: DependencyDirection
) -> impl Iterator<Item = FeatureMetadata<'g>> + 'a

Returns the set of "root feature" metadatas in the specified direction.

If direction is Forward, return the set of metadatas that do not have any dependencies within the selected graph.

If direction is Reverse, return the set of metadatas that do not have any dependents within the selected graph.

Cycles

If a root consists of a dependency cycle, all the packages in it will be returned in arbitrary order.}

pub fn cross_links<'a>(
&'a self,
direction: DependencyDirection
) -> impl Iterator<Item = CrossLink<'g>> + 'a

Creates an iterator over CrossLink instances in the direction specified.

Cycles

The links in a dependency cycle may be returned in arbitrary order.