Trait guppy::graph::feature::FeatureFilter

pub trait FeatureFilter {
    fn accept(&mut self, graph: &FeatureGraph, feature_id: FeatureId) -> bool;
}

Trait representing whether a feature within a package should be selected.

This is conceptually similar to passing --features or other similar command-line options to Cargo.

Most uses will involve using one of the predefined filters: all_filter, default_filter, or none_filter. For advanced uses, the trait is implemented for all functions that match for<'g> FnMut(&FeatureGraph<'g>, FeatureId<'g>) -> bool. The filter_fn helper is provided to assist with type inference.

Required methods

fn accept(&mut self, graph: &FeatureGraph, feature_id: FeatureId) -> bool

Returns true if this feature ID should be selected in the graph.

Returning false does not prevent this feature ID from being included if it's reachable through other means.

In general, accept should return true if feature_id.is_base() is true.

The feature ID is guaranteed to be in this graph, so it is OK to panic if it isn't found.

Loading content...