Trait guppy::graph::feature::FeatureResolver

pub trait FeatureResolver<'g> {
    fn accept(
        &mut self, 
        query: &FeatureQuery<'g>, 
        link: ConditionalLink<'g>
    ) -> bool;
}

Represents whether a particular link within a feature graph should be followed during a resolve operation.

Required methods

fn accept(
&mut self,
query: &FeatureQuery<'g>,
link: ConditionalLink<'g>
) -> bool

Returns true if this conditional link should be followed during a resolve operation.