Trait guppy::graph::feature::FeatureResolver

pub trait FeatureResolver<'g> {
    fn accept(&mut self, query: &FeatureQuery<'g>, link: CrossLink<'g>) -> bool;
}

Represents whether a particular link within a feature graph should be followed during a resolve operation.

Required methods

fn accept(&mut self, query: &FeatureQuery<'g>, link: CrossLink<'g>) -> bool

Returns true if this cross-package link should be followed during a resolve operation.