guppy::graph::feature
Trait FeatureResolver

pub trait FeatureResolver<'g> {
    // Required method
    fn accept(
        &mut self,
        query: &FeatureQuery<'g>,
        link: ConditionalLink<'g>,
    ) -> bool;
}

Represents whether a particular link within a feature graph should be followed during a resolve operation.

Required Methods

fn accept( &mut self, query: &FeatureQuery<'g>, link: ConditionalLink<'g>, ) -> bool

Returns true if this conditional link should be followed during a resolve operation.