Struct guppy::graph::PackageGraph

pub struct PackageGraph { /* fields omitted */ }

A graph of packages and dependencies between them, parsed from metadata returned by cargo metadata.

For examples on how to use PackageGraph, see the examples directory in this crate.

Methods

impl PackageGraph

pub fn from_command(command: &mut MetadataCommand) -> Result<Self, Error>

Constructs a package graph from the given command.

pub fn from_json(json: impl AsRef<str>) -> Result<Self, Error>

Constructs a package graph from the given JSON output of cargo metadata.

pub fn new(metadata: Metadata) -> Result<Self, Error>

Constructs a package graph from the given metadata.

pub fn workspace(&self) -> &Workspace

Returns information about the workspace.

pub fn package_ids(
&self
) -> impl Iterator<Item = &PackageId> + ExactSizeIterator

Returns an iterator over all the package IDs in this graph.

pub fn packages(
&self
) -> impl Iterator<Item = &PackageMetadata> + ExactSizeIterator

Returns an iterator over all the packages in this graph.

pub fn package_count(&self) -> usize

Returns the number of packages in this graph.

pub fn link_count(&self) -> usize

Returns the number of links in this graph.

pub fn metadata(&self, package_id: &PackageId) -> Option<&PackageMetadata>

Returns the metadata for the given package ID.

pub fn retain_edges<F>(&mut self, visit: F) where
F: Fn(&PackageGraphData, DependencyLink) -> bool,

Keeps all edges that return true from the visit closure, and removes the others.

The order edges are visited is not specified.

pub fn new_depends_cache(&self) -> DependsCache

Creates a new cache for depends_on queries.

The cache is optional but can speed up some queries.

pub fn depends_on(
&self,
package_a: &PackageId,
package_b: &PackageId
) -> Result<bool, Error>

Returns true if package_a depends (directly or indirectly) on package_b.

In other words, this returns true if package_b is a (possibly transitive) dependency of package_a.

For repeated queries, consider using new_depends_cache to speed up queries.

pub fn cycles(&self) -> Cycles

Returns information about dependency cycles in this graph.

For more information, see the documentation for Cycles.

pub fn dep_links_directed<'g>(
&'g self,
package_id: &PackageId,
dep_direction: DependencyDirection
) -> Option<impl Iterator<Item = DependencyLink<'g>> + 'g>

Returns the direct dependencies for the given package ID in the specified direction.

pub fn dep_links<'g>(
&'g self,
package_id: &PackageId
) -> Option<impl Iterator<Item = DependencyLink<'g>> + 'g>

Returns the direct dependencies for the given package ID.

pub fn reverse_dep_links<'g>(
&'g self,
package_id: &PackageId
) -> Option<impl Iterator<Item = DependencyLink<'g>> + 'g>

Returns the direct reverse dependencies for the given package ID.

impl PackageGraph

Helpers for property testing

The methods in this section allow a PackageGraph to be used in property-based testing scenarios.

Currently, proptest 0.9 is supported if the proptest09 feature is enabled.

pub fn prop09_id_strategy<'g>(
&'g self
) -> impl Strategy<Value = &'g PackageId> + 'g

Returns a Strategy that generates random package IDs from this graph.

Requires the proptest09 feature to be enabled.

Panics

Panics if there are no packages in this PackageGraph.

pub fn prop09_link_strategy<'g>(
&'g self
) -> impl Strategy<Value = DependencyLink<'g>> + 'g

Returns a Strategy that generates random dependency links from this graph.

Requires the proptest09 feature to be enabled.

Panics

Panics if there are no dependency edges in this PackageGraph.

pub fn prop09_resolver_strategy<'g>(
&'g self
) -> impl Strategy<Value = Prop09Resolver> + 'g

Returns a Strategy that generates a random PackageResolver instance from this graph.

Requires the proptest09 feature to be enabled.

impl PackageGraph

Selectors

The methods in this section create package selectors, which are queries over subsets of this package graph. Use the methods here for queries based on transitive dependencies.

pub fn select_workspace(&self) -> PackageSelect

Creates a new forward selector over the entire workspace.

select_workspace will select all workspace packages and their transitive dependencies.

pub fn select_all(&self) -> PackageSelect

Creates a new selector that returns all members of this package graph.

This is normally the same as select_workspace, but can differ in some cases:

if packages have been replaced with [patch] or [replace]

if some edges have been removed from this graph with retain_edges.

In most situations, select_workspace is preferred. Use select_all if you know you need parts of the graph that aren't accessible from the workspace.

pub fn select_directed<'g, 'a>(
&'g self,
package_ids: impl IntoIterator<Item = &'a PackageId>,
dep_direction: DependencyDirection
) -> Result<PackageSelect<'g>, Error>

Creates a new selector that returns transitive dependencies of the given packages in the specified direction.

Returns an error if any package IDs are unknown.

pub fn select_transitive_deps_directed<'g, 'a>(
&'g self,
package_ids: impl IntoIterator<Item = &'a PackageId>,
dep_direction: DependencyDirection
) -> Result<PackageSelect<'g>, Error>

Deprecated since 0.1.3:
renamed to select_directed

Creates a new selector that returns transitive dependencies of the given packages in the specified direction.

Returns an error if any package IDs are unknown.

pub fn select_forward<'g, 'a>(
&'g self,
package_ids: impl IntoIterator<Item = &'a PackageId>
) -> Result<PackageSelect<'g>, Error>

Creates a new selector that returns transitive dependencies of the given packages.

Returns an error if any package IDs are unknown.

pub fn select_transitive_deps<'g, 'a>(
&'g self,
package_ids: impl IntoIterator<Item = &'a PackageId>
) -> Result<PackageSelect<'g>, Error>

Deprecated since 0.1.3:
renamed to select_forward

Creates a new selector that returns transitive dependencies of the given packages.

Returns an error if any package IDs are unknown.

pub fn select_reverse<'g, 'a>(
&'g self,
package_ids: impl IntoIterator<Item = &'a PackageId>
) -> Result<PackageSelect<'g>, Error>

Creates a new selector that returns transitive reverse dependencies of the given packages.

Returns an error if any package IDs are unknown.

pub fn select_transitive_reverse_deps<'g, 'a>(
&'g self,
package_ids: impl IntoIterator<Item = &'a PackageId>
) -> Result<PackageSelect<'g>, Error>

Deprecated since 0.1.3:
renamed to select_reverse

Creates a new selector that returns reverse transitive dependencies of the given packages.

Returns an error if any package IDs are unknown.