Struct guppy::graph::PackageSet

pub struct PackageSet<'g> { /* private fields */ }

A set of resolved packages in a package graph.

Created by PackageQuery::resolve.

Implementations

impl<'g> PackageSet<'g>

pub fn len(&self) -> usize

Returns the number of packages in this set.

pub fn is_empty(&self) -> bool

Returns true if no packages were resolved in this set.

pub fn contains(&self, package_id: &PackageId) -> Result<bool, Error>

Returns true if this package ID is contained in this resolve set.

Returns an error if the package ID is unknown.

pub fn to_package_query( &self, direction: DependencyDirection ) -> PackageQuery<'g>

Creates a new PackageQuery from this set in the specified direction.

This is equivalent to constructing a query from all the package_ids.

pub fn union(&self, other: &Self) -> Self

Returns a PackageSet that contains all packages present in at least one of self and other.

Panics

Panics if the package graphs associated with self and other don’t match.

pub fn intersection(&self, other: &Self) -> Self

Returns a PackageSet that contains all packages present in both self and other.

Panics

Panics if the package graphs associated with self and other don’t match.

pub fn difference(&self, other: &Self) -> Self

Returns a PackageSet that contains all packages present in self but not other.

Panics

Panics if the package graphs associated with self and other don’t match.

pub fn symmetric_difference(&self, other: &Self) -> Self

Returns a PackageSet that contains all packages present in exactly one of self and other.

Panics

Panics if the package graphs associated with self and other don’t match.

pub fn filter( &self, direction: DependencyDirection, callback: impl FnMut(PackageMetadata<'g>) -> bool ) -> Self

Returns a PackageSet on which a filter has been applied.

Filters out all values for which the callback returns false.

Cycles

For packages within a dependency cycle, the callback will be called in non-dev order. When the direction is forward, if package Foo has a dependency on Bar, and Bar has a cyclic dev-dependency on Foo, then Foo is returned before Bar.

pub fn partition( &self, direction: DependencyDirection, callback: impl FnMut(PackageMetadata<'g>) -> bool ) -> (Self, Self)

Partitions this PackageSet into two.

The first PackageSet contains packages for which the callback returned true, and the second one contains packages for which the callback returned false.

Cycles

For packages within a dependency cycle, the callback will be called in non-dev order. When the direction is forward, if package Foo has a dependency on Bar, and Bar has a cyclic dev-dependency on Foo, then Foo is returned before Bar.

pub fn filter_partition( &self, direction: DependencyDirection, callback: impl FnMut(PackageMetadata<'g>) -> Option<bool> ) -> (Self, Self)

Performs filtering and partitioning at the same time.

The first PackageSet contains packages for which the callback returned Some(true), and the second one contains packages for which the callback returned Some(false). Packages for which the callback returned None are dropped.

Cycles

For packages within a dependency cycle, the callback will be called in non-dev order. When the direction is forward, if package Foo has a dependency on Bar, and Bar has a cyclic dev-dependency on Foo, then Foo is returned before Bar.

pub fn to_feature_set(&self, filter: impl FeatureFilter<'g>) -> FeatureSet<'g>

Creates a new FeatureSet consisting of all packages in this PackageSet, using the given feature filter.

This will cause the feature graph to be constructed if it hasn’t been done so already.

pub fn package_ids<'a>( &'a self, direction: DependencyDirection ) -> impl Iterator<Item = &'g PackageId> + ExactSizeIterator + 'a

Iterates over package IDs, in topological order in the direction specified.

Cycles

The packages within a dependency cycle will be returned in non-dev order. When the direction is forward, if package Foo has a dependency on Bar, and Bar has a cyclic dev-dependency on Foo, then Foo is returned before Bar.

pub fn packages<'a>( &'a self, direction: DependencyDirection ) -> impl Iterator<Item = PackageMetadata<'g>> + ExactSizeIterator + 'a

Iterates over package metadatas, in topological order in the direction specified.

Cycles

The packages within a dependency cycle will be returned in non-dev order. When the direction is forward, if package Foo has a dependency on Bar, and Bar has a cyclic dev-dependency on Foo, then Foo is returned before Bar.

pub fn root_ids<'a>( &'a self, direction: DependencyDirection ) -> impl Iterator<Item = &'g PackageId> + ExactSizeIterator + 'a

Returns the set of “root package” IDs in the specified direction.

If direction is Forward, return the set of packages that do not have any dependencies within the selected graph.

If direction is Reverse, return the set of packages that do not have any dependents within the selected graph.

Cycles

If a root consists of a dependency cycle, all the packages in it will be returned in non-dev order (when the direction is forward).

pub fn root_packages<'a>( &'a self, direction: DependencyDirection ) -> impl Iterator<Item = PackageMetadata<'g>> + ExactSizeIterator + 'a

Returns the set of “root package” metadatas in the specified direction.

If direction is Forward, return the set of packages that do not have any dependencies within the selected graph.

If direction is Reverse, return the set of packages that do not have any dependents within the selected graph.

Cycles

If a root consists of a dependency cycle, all the packages in it will be returned in non-dev order (when the direction is forward).

pub fn links<'a>( &'a self, direction: DependencyDirection ) -> impl Iterator<Item = PackageLink<'g>> + 'a

Creates an iterator over PackageLink instances.

If the iteration is in forward order, for any given package, at least one link where the package is on the to end is returned before any links where the package is on the from end.

If the iteration is in reverse order, for any given package, at least one link where the package is on the from end is returned before any links where the package is on the to end.

Cycles

The links in a dependency cycle will be returned in non-dev order. When the direction is forward, if package Foo has a dependency on Bar, and Bar has a cyclic dev-dependency on Foo, then the link Foo -> Bar is returned before the link Bar -> Foo.

pub fn display_dot<'a, V: PackageDotVisitor + 'g>( &'a self, visitor: V ) -> impl Display + 'a

Constructs a representation of the selected packages in dot format.