Module guppy::graph::summaries

Generate build summaries from CargoSet instances.

Requires the summaries feature to be enabled.

Modules

diff
Compare and diff summaries.

Structs

CargoOptionsSummary
A summary of Cargo options used to build a CargoSet.

PackageInfo
Information about a package in a summary that isn't part of the unique identifier.

SummaryId
A unique identifier for a package in a build summary.

SummaryWithMetadata
An in-memory representation of a build summary.

Enums

PackageStatus
The status of a package in a summary, such as whether it is part of the initial build set.

SummarySource
The location of a package.

Type Definitions

PackageMap
A type representing a package map as used in Summary instances.

Summary
A type alias for build summaries generated by guppy.