Struct guppy::graph::summaries::SummaryWithMetadata

pub struct SummaryWithMetadata<M = Value> {
    pub metadata: Option<M>,
    pub target_packages: BTreeMap<SummaryId, PackageInfo>,
    pub host_packages: BTreeMap<SummaryId, PackageInfo>,
}

An in-memory representation of a build summary.

The metadata parameter is customizable.

For more, see the crate-level documentation.

Fields
metadata: Option<M>

Extra metadata associated with the summary.

This may be used for storing extra information about the summary.

The type defaults to toml::Value but is customizable.

target_packages: BTreeMap<SummaryId, PackageInfo>

The packages and features built on the target platform.

host_packages: BTreeMap<SummaryId, PackageInfo>

The packages and features built on the host platform.

Implementations

impl<M> SummaryWithMetadata<M>

pub fn parse<'de>(s: &'de str) -> Result<SummaryWithMetadata<M>, Error> where
M: Deserialize<'de>,

Deserializes a summary from the given string, with optional custom metadata.

pub fn diff<M2>(&'a self, other: &'a SummaryWithMetadata<M2>) -> SummaryDiff<'a>

Perform a diff of this summary against another.

This doesn't diff the metadata, just the initials and packages.

pub fn to_string(&self) -> Result<String, Error> where
M: Serialize,

Serializes this summary to a TOML string.

pub fn write_to_string(&self, dst: &mut String) -> Result<(), Error> where
M: Serialize,

Serializes this summary into the given TOML string, using pretty TOML syntax.