Trait guppy::graph::PackageDotVisitor

pub trait PackageDotVisitor {
    fn visit_package(
        &self,
        package: PackageMetadata<'_>,
        f: &mut DotWrite<'_, '_>
    ) -> Result;
    fn visit_link(
        &self,
        link: PackageLink<'_>,
        f: &mut DotWrite<'_, '_>
    ) -> Result;
}

A visitor used for formatting dot graphs.

Required Methods

fn visit_package(
&self,
package: PackageMetadata<'_>,
f: &mut DotWrite<'_, '_>
) -> Result

Visits this package. The implementation may output a label for this package to the given DotWrite.

fn visit_link(&self, link: PackageLink<'_>, f: &mut DotWrite<'_, '_>) -> Result

Visits this dependency link. The implementation may output a label for this link to the given DotWrite.