Trait guppy::graph::PackageResolver

pub trait PackageResolver<'g> {
    fn accept(
        &mut self, 
        query: &PackageQuery<'g>, 
        link: PackageLink<'g>
    ) -> bool;
}

Represents whether a particular link within a package graph should be followed during a resolve operation.

Required Methods

fn accept(&mut self, query: &PackageQuery<'g>, link: PackageLink<'g>) -> bool

Returns true if this link should be followed during a resolve operation.

Returning false does not prevent the to package (or from package with query_reverse) from being included if it’s reachable through other means.